resize imagenes

+++ src/Window/Position.h
@@ -3,18 +3,18 @@
 
 class Position{
 	private:
-		int x;
-		int y;
+		float x;
+		float y;
 
 	public:
-		Position(int x, int y);
+		Position(float x, float y);
 		~Position();
 
 		bool operator==(const Position& other);
 
-		int getX() const;
+		float getX() const;
 
-        int getY() const;
+        float getY() const;
 
 };
